ShopEaseDocs/requirements.txt
Django>=5.0
python-decouple>=3.8
python-dotenv>=1.0.0
mysqlclient>=2.2.0
Pillow>=10.0.0
